+++ src/Client.php
@@ -3,7 +3,11 @@
 namespace MarketDataApp;
 
 use MarketDataApp\Endpoints\Indices;
+use MarketDataApp\Endpoints\Markets;
+use MarketDataApp\Endpoints\MutualFunds;
+use MarketDataApp\Endpoints\Options;
 use MarketDataApp\Endpoints\Stocks;
+use MarketDataApp\Endpoints\Utilities;
 
 class Client extends ClientBase
 {
@@ -19,11 +23,39 @@ class Client extends ClientBase
      */
     public Stocks $stocks;
 
+    /**
+     * The Market Data API provides a comprehensive suite of options endpoints, designed to cater to various needs
+     * around options data. These endpoints are designed to be flexible and robust, supporting both real-time
+     * and historical data queries. They accommodate a wide range of optional parameters for detailed data
+     * retrieval, making the Market Data API a versatile tool for options traders and financial analysts.
+     */
+    public Options $options;
+
+    /**
+     * The Markets endpoints provide reference and status data about the markets covered by Market Data.
+     */
+    public Markets $markets;
+
+    /**
+     * The mutual funds endpoints offer access to historical pricing data for mutual funds.
+     */
+    public MutualFunds $mutual_funds;
+
+    /**
+     * These endpoints are designed to assist with API-related service issues, including checking the online status and
+     * uptime.
+     */
+    public Utilities $utilities;
+
     public function __construct($token)
     {
         parent::__construct($token);
 
         $this->indices = new Indices($this);
         $this->stocks = new Stocks($this);
+        $this->options = new Options($this);
+        $this->markets = new Markets($this);
+        $this->mutual_funds = new MutualFunds($this);
+        $this->utilities = new Utilities($this);
     }
 }