+++ src/Endpoints/Responses/Markets/Status.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace MarketDataApp\Endpoints\Responses\Markets;
+
+class Status
+{
+}