+++ src/Endpoints/Responses/MutualFunds/Candles.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace MarketDataApp\Endpoints\Responses\MutualFunds;
+
+class Candles
+{
+}