+++ src/Endpoints/Responses/Options/OptionChain.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace MarketDataApp\Endpoints\Responses\Options;
+
+class OptionChain
+{
+}