+++ src/Endpoints/Responses/Utilities/ApiStatus.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace MarketDataApp\Endpoints\Responses\Utilities;
+
+class ApiStatus
+{
+}